add ui classes for taskgroup editor

+++ core/database/src/main/kotlin/net/onefivefour/sessiontimer/core/database/data/DatabaseDefaultValues.kt
@@ -3,7 +3,7 @@ package net.onefivefour.sessiontimer.core.database.data
 interface DatabaseDefaultValues {
     fun getSessionTitle(): String
     fun getTaskGroupTitle(): String
-    fun getTaskGroupColor(): Int
+    fun getTaskGroupColor(): Long
     fun getTaskTitle(): String
     fun getTaskDuration(): Int
 }
